16. Filter and sort datable of fish catches
We frame the results of the filters with a turbo frame `<%=
turbo_frame_tag 'fish_catches' do %>`. This way, all links inside the
frame (i.e sort links) are updating the results without loading the HTML
of the whole page.

However, when submitting the form, we still receive the HTML of the
whole page. In order to update only the catches, we need to target the
results. If we extend the frame and include the filters as well, we will
constantly lose focus from the form once it's submitted. This happens
because the frame of the response will include the filter's form as well
so it will be updated. So we need to exclude the filters from the
updated frame. In other words, the form should target the frame. This is
possible by adding the data attribute `data-turbo-frame` to the form.
This way, the filters don't change because they are out of the frame.

In addition, the existing Stimulus form controller is reused in order to
submit the form automatically once some input is given. The submit
button is removed as well.

We also add the data-turbo-action='advance' in order to update the
browser history.

Last, we add data-turbo-frame="top" for the link "Log new catch" in
order to break out of the frame and load the new catch form. If we don't
add it, the framework falls back to full page reload.