17. Update hidden fields of filters form with sorting criteria
Here we update the hidden fields of the filters forms on the fish
catches page with JS in order to have the same sort criteria and
direction as given by the user when interacting with the sort links.

We create a new stimulus controller and connect it with the whole page
of the fish_catches/index (see `<div data-controller="sort-link">`). We
define two stimulus targets in order to have easier element handling in
the stimulus handling. And we handle every click on a sort link with a
controller action. Note that we use the `turbo:click` as event instead
of `click` because the sort links are inside a turbo frame and the
clicks on them are intercepted.